Ensure no trailing back-slashes on deployment URLs
This commit updates the `get_maybe_local_path` helper to strip trailing
back-slashes from input paths.

This is primarily useful when a cloud storage bucket is passed as a URL,
to make sure empty `/` directories are not created in the full path.

Deployments are updated to use this helper when constructing their full
URLs.