Make slurm execute script customizable
This is done by making the template source path pointing to a
workflow_manager_variable.

The major change is around extending `register_template`'s `src_path`
prop (renamed from `src_name`) to allow it to point to a user-specified
path.

Another change is to point the destination of the slurm execute script
to `execute_experiment`. The intent is to make the ux similar to
workspaces that don't use workflow manager, in that manually changing
the rendered `execute_experiment` in the experiment_run_dir takes
effect.