Create a user-managed package manager
This commit creates a user-managed package manager that behaves as the
previous None package manager was.

Additionally, the None package manager has been converted to not do
anything. Paths are no longer required to be defined when using a None
package manager anymore. The default package manager is updated to be
`user-managed`.